Separate transmitter GNSS functionality into its own source file
