+++ app_radio_module/src/transmitter_gnss.h
@@ -1,0 +1,6 @@
+#ifndef RADIO_MODULE_TRANSMITTER_GNSS_H
+#define RADIO_MODULE_TRANSMITTER_GNSS_H
+
+void init_gnss(void);
+
+#endif //RADIO_MODULE_TRANSMITTER_GNSS_H
