+++ app_radio_module/src/transmitter_smf.h
@@ -1,0 +1,10 @@
+#ifndef RADIO_MODULE_TRANSMITTER_SMF_H
+#define RADIO_MODULE_TRANSMITTER_SMF_H
+
+typedef enum {
+    GROUND_STATE = 0,
+    FLIGHT_STATE,
+    DEBUG_STATE
+} TRANSMITTER_STATES;
+
+#endif //RADIO_MODULE_TRANSMITTER_SMF_H
